Make neo4j-graph-algorithms work with 3.2 and all the Kernel-API changes
This creates a new "Kernel" class that wraps the API to make it behave like the old one
Going forward we should rather change the code to not rely on Cursors during the loading
but just use primitive values and iterators.

Something that is also worse now in the new Kernel API is the massive use of (checked) Exceptions
which is definitely not performance improving.